feat: implement affiliates management and enhance anonymity handling
- Introduced a new Affiliates page for creators to review performance across all affiliates promoting their projects.
- Added detailed views for individual affiliates, including performance metrics and testimonials.
- Enhanced UI to handle anonymous names for both marketers and projects, displaying appropriate fallback content.
- Updated navigation links and sidebar to reflect the new affiliates structure, improving user experience in the creator dashboard.
- Refactored components to ensure consistent handling of anonymous entities across the application.
